Replacing all BigIntegers with int or double, replacing all ConcurrentDictionaries through Dictionary, introducing DictionaryExtention to provide missing Dictionary functionality
git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@7128 9b0d660e-8ab2-dd11-ad6d-000c29d71a56

+++ CrypPluginsExperimental/TestVectorGenerator/TestVectorGenerator.cs
@@ -158,7 +158,7 @@ namespace Cryptool.Plugins.TestVectorGenerator
                 _currentTextLength = _settings.TextLength;
 
             _startSentence = _rand.Next(0, _inputArray.Length);
-            BigInteger count = 0;
+            int count = 0;
             while (_plaintextList.Exists(s => s.StartsWith(_inputArray[_startSentence])))
             {
                 _startSentence = _rand.Next(0, _inputArray.Length);
